+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [1.2.0] - 2023-10-03
 - Updated to work in Elasticsearch v8.5.2
+- Replaced the arg_parser code with gen_class.ArgParser code.
 
 ### Changed
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
 - main, run_program: Removed gen_libs.get_inst call.
 - Documentation update